moved anims into own package, reusable
